+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Unlockable door blocks.
 - Sprite rotation for Spots.
 - Better entity placement options for spots.
+- Camera zoom can now be locked, to prevent accidental zooming while rolling in combat. It comes
+  with a keybind to enable/disable the setting, and an Auto/Toggle behavior setting. Auto behavior
+  will only lock the camera zoom while movement and combat inputs are also being pressed.
 
 ### Changed
 - Bats move slower and use a simple proportional controller to maintain altitude